move BaseModelWorker outside serve.model_worker to make it independent of torch and etc. So user can create custom model worker besides local models, such as wrap a online api to model worker whithout torch and transformers.
