+++ fastchat/serve/huggingface_api_worker.py
@@ -28,7 +28,7 @@ from fastapi.responses import JSONResponse, StreamingResponse
 from huggingface_hub import InferenceClient
 
 from fastchat.constants import SERVER_ERROR_MSG, ErrorCode
-from fastchat.serve.model_worker import BaseModelWorker
+from fastchat.serve.base_model_worker import BaseModelWorker
 from fastchat.utils import build_logger
 
 worker_id = str(uuid.uuid4())[:8]
